requirements.txt
requests-toolbelt
requests
pyinstaller
pycryptodomex
cryptography
tk
pillow
pypiwin32